% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_rule.R
\name{add_sent_list}
\alias{add_sent_list}
\title{Add a dummy columm}
\usage{
add_sent_list(pos_tibble)
}
\arguments{
\item{pos_tibble}{A tibble.}
}
\value{
\code{tibble} with a column of 1
}
\description{
Add a dummy columm
}
