% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_basic.R
\name{as_basic}
\alias{as_basic}
\title{Convert Tags to Basic Tags}
\usage{
as_basic(x, ...)
}
\arguments{
\item{x}{A \code{tag_pos} object or a named list of vectors.}

\item{\ldots}{ignored.}
}
\value{
Returns a combined character vector of words and universal tags.
}
\description{
Convert the Penn Treebank tags to basic part of speech tags.
}
\details{
Basic tags include: (a) nouns, (b) adjectives, (c) prepositions,
(d) articles, (e) verbs, (f) pronouns, (g) adverbs, (h) interjections, &
(i) conjunctions.  The \code{X} and \code{.} tags are retained for punctuation
and unclassified parts of speech.  This tagset can be useful for more coarse
purposes, including formality (Heylighen & Dewaele, 2002) scoring.
}
\examples{
(x <- tag_pos("They refuse to permit us to obtain the refuse permit"))
as_basic(x)

(out1 <- tag_pos(sam_i_am))
as_basic(out1)

presidential_debates_2012_pos
as_basic(presidential_debates_2012_pos)

library(dplyr)
presidential_debates_2012_pos \%>\%
    as_basic() \%>\%
    count_tags(grouping.var=presidential_debates_2012[["person"]]) \%>\%
    plot()
}
\references{
Heylighen, F., & Dewaele, J.M. (2002). Variation in the
contextuality of language: An empirical measure. Context in Context, Special
issue of Foundations of Science, 7 (3), 293-340.
}
\keyword{explicit,}
\keyword{formality,}
\keyword{parts-of-speech,}
\keyword{pos}
