% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_pos.R
\name{as_pos}
\alias{as_pos}
\alias{as_tokens}
\title{Extract Parts of Speech or Tokens from a \code{tag_pos} Object}
\usage{
as_pos(x, collapse = TRUE, ...)

as_tokens(x, collapse = TRUE, ...)
}
\arguments{
\item{x}{A \code{\link[tagger]{tag_pos}} object.}

\item{collapse}{logical.  If \code{TRUE} the pos/tokens will be collapsed
into a vector of paste collapsed strings.  If \code{FALSE} the pos/tokens
will be returned as a list of vectors.}

\item{\ldots}{ignored.}
}
\value{
Returns a vector of strings or a list of pos/tokens.
}
\description{
Extract parts of speech or corresponding tokens from a
\code{\link[tagger]{tag_pos}} object.  This is useful for extracting
part of speech based text features for modeling.
}
\examples{
library(dplyr)
presidential_debates_2012_pos \%>\%
    select_tags(c("NN", "NNP", "NNPS", "NNS")) \%>\%
    as_pos()

presidential_debates_2012_pos \%>\%
    select_tags(c("NN", "NNP", "NNPS", "NNS")) \%>\%
    as_pos(collapse = FALSE)


presidential_debates_2012_pos \%>\%
    select_tags(c("NN", "NNP", "NNPS", "NNS")) \%>\%
    as_tokens()

presidential_debates_2012_pos \%>\%
    select_tags(c("NN", "NNP", "NNPS", "NNS")) \%>\%
    as_tokens(collapse = FALSE)
}
