% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as_word_tag.R
\name{as_word_tag}
\alias{as_word_tag}
\title{Combine Words + Tags}
\usage{
as_word_tag(x, sep = "/", word.first = TRUE, ...)
}
\arguments{
\item{x}{A \code{tag_pos} object or a named list of vectors.}

\item{sep}{A separator to put between the word and the tag.}

\item{word.first}{logical.  If \code{TRUE} the word comes before the tag.}

\item{\ldots}{ignored.}
}
\value{
Returns a combined character vector of words and tags.
}
\description{
A traditional style of displaying words and tags in the same vector.
}
\examples{
(x <- tag_pos("I need $54 to go to the movies."))
c(x) ## The true structure of a `tag_pos` object
as_word_tag(c(x))
as_word_tag(c(x), word.first=FALSE)
as_word_tag(c(x), sep="|")
gsub("([^ ]+)", "(\\\\1)", as_word_tag(c(x), sep="="))

(out1 <- tag_pos(sam_i_am))
c(out1)
as_word_tag(c(out1))
}

