% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{binary_levels}
\alias{binary_levels}
\title{Build binary clustering levels up to k}
\usage{
binary_levels(k_deep)
}
\arguments{
\item{k_deep}{Desired number of clusters at the deepest level (this must be a power of 2).}
}
\value{
Integer vector like c(k, k/2, ..., 2) for use as clusters_by_level.
Level 1 = most granular (k clusters), last level = 2 clusters.
}
\description{
Build binary clustering levels up to k
}
\examples{
binary_levels(8)
}
