% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{canon_tag}
\alias{canon_tag}
\title{Canonicalize Tags for Deterministic Comparison}
\usage{
canon_tag(x)
}
\arguments{
\item{x}{Tag to normalize}
}
\description{
This function normalizes tags as the first step in the tag cleanup pipeline.
The normalization steps are as follows:  remove NA's, remove whitespace, check
that the tags use only lowercase ASCII characters, any separators are unified
to spaces.
}
