% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stanford.R
\name{check_stanford_installed}
\alias{check_stanford_installed}
\title{Check if Stanford \pkg{CoreNLP} is Installed and In Root}
\usage{
check_stanford_installed(stanford = tagger::coreNLP_loc(),
  download = tagger::coreNLP_url(), verbose = TRUE)
}
\arguments{
\item{stanford}{The version of Stanford's \pkg{CoreNLP}.}

\item{download}{The download url for Stanford's \pkg{CoreNLP}.}

\item{verbose}{If \code{TRUE} messages are printed even when everything is
installed.}
}
\description{
Checks that Stanford \pkg{CoreNLP} is installed and in root.
}
\examples{
\dontrun{
check_stanford_installed()
}
}
\references{
http://nlp.stanford.edu/software/corenlp.shtml
}
\keyword{stanford}

