% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{clean_tag_matrix}
\alias{clean_tag_matrix}
\title{Clean a given Tag Matrix}
\usage{
clean_tag_matrix(tag_matrix, tag_cols = NULL)
}
\arguments{
\item{tag_matrix}{The full generated tag matrix as produced by the tagging pipeline}

\item{tag_cols}{Cluster tags at consecutive levels (level 1 being the most generic).}
}
\description{
After the LLM tagging pipeline has been executed this function is used to audit
the generated tags matrix. Removing duplicates and checking the tag structure.
}
