% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{clean_tag_row}
\alias{clean_tag_row}
\title{Clean Tag Row}
\usage{
clean_tag_row(tags, tag_cols)
}
\arguments{
\item{tags}{Raw tag vector of length k (number of clustering levels).}

\item{tag_cols}{Cluster tags at consecutive levels. (Level 1 being most generic)}
}
\description{
This function is called for each row in the tag matrix. With each iteration
(row) the raw tags are first normalized and then consecutive duplicates are
removed. After duplicates are collapsed lower levels are padded with NA's.
}
