% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_embeddings}
\alias{cluster_embeddings}
\title{Cluster embeddings and return hclust object}
\usage{
cluster_embeddings(embeddings, clusters_by_level)
}
\arguments{
\item{embeddings}{List or matrix of embeddings.}

\item{clusters_by_level}{Integer vector of cluster counts by level.}
}
\value{
A list with hclust and distance matrix.
}
\description{
Cluster embeddings and return hclust object
}
\examples{
embeddings <- matrix(rnorm(40), ncol = 4)
cluster_embeddings(embeddings, clusters_by_level = c(4, 2))
}
