% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{collapse_consecutive}
\alias{collapse_consecutive}
\title{Collapse Consecutive Duplicate Tags}
\usage{
collapse_consecutive(tags)
}
\arguments{
\item{tags}{A character vector representing the final tag list.}
}
\description{
This function collapses repeat tags that appear over consecutive levels.
}
