% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/count_pos.R
\name{count_pos}
\alias{count_pos}
\title{Counts of Part of Speech Tags by Grouping Variable(s)}
\usage{
count_pos(x, grouping.var = NULL, group.names, pretty = TRUE, ...)
}
\arguments{
\item{x}{A \code{tag_pos} object or a named list of vectors.}

\item{grouping.var}{The grouping variable(s).  Default \code{NULL} generates
one row for each element in \code{x}.}

\item{group.names}{A vector of names that corresponds to group.var.  Generally
for internal use.}

\item{pretty}{logical.  If \code{TRUE} pretty printing is used.  Pretty
printing can be turned off globally by setting
\code{options(tagger_pretty = FALSE)}.}

\item{\ldots}{ignored}
}
\value{
Returns a \code{\link[dplyr]{tbl_df}} of counts of parts of speech.
}
\description{
Tabulate counts of part of speech tags by grouping variable(s).
}
\examples{
data(presidential_debates_2012_pos) # pre tagged data
count_pos(presidential_debates_2012_pos)
count_pos(presidential_debates_2012_pos, by = presidential_debates_2012$person)
with(presidential_debates_2012,
    count_pos(presidential_debates_2012_pos, by = list(person, time))
)
}
\keyword{count}
\keyword{proportion}

