% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionLevel.R
\name{extract_closed_answer_entities}
\alias{extract_closed_answer_entities}
\title{Extract Closed Answer entities from the Codebook.}
\usage{
extract_closed_answer_entities(codebook, base_uri, procedure_uri)
}
\arguments{
\item{codebook}{The survey codebook as a dataframe.}

\item{base_uri}{The base URI string to prepend to generated IDs.}

\item{procedure_uri}{The full IRI of the sur:SurveyProcedure these answers belong to.}
}
\value{
A list of R lists, where each inner list represents a sur:ClosedAnswer entity.
}
\description{
This function extracts predefined answer options for closed questions
and formats them as sur:ClosedAnswer entities.
}
