% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submissionLevel.R
\name{extract_participant_entities}
\alias{extract_participant_entities}
\title{Extract Fieldworker Entities}
\usage{
extract_participant_entities(main, base_uri)
}
\arguments{
\item{main}{The main data frame of a survey where all the submissions are stored.}

\item{base_uri}{The base URI to be prepended to the generated IDs.}
}
\value{
A list of R lists, where each inner list represents a sur:Participant entity.
}
\description{
This function extracts fieldworkers entities.
}
