% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionLevel.R
\name{extract_question_entities}
\alias{extract_question_entities}
\title{Extract Question and Talk entities from the questions dataframe.}
\usage{
extract_question_entities(
  questions,
  base_uri,
  procedure_uri,
  question_classifier,
  tag_set = new.env()
)
}
\arguments{
\item{questions}{The survey questions dataframe.}

\item{base_uri}{The base URI string to prepend to any generated IDs.}

\item{procedure_uri}{The full IRI of the sur:SurveyProcedure these questions belong to.}
}
\value{
A list of R lists, where each inner list represents a sur:Question or sur:Talk entity.
}
\description{
This function formats survey questions according to the Survey Ontology,
mapping question types to appropriate classes. These questions are then
grouped by section for tagging.
}
