% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submissionLevel.R
\name{extract_submission_entities}
\alias{extract_submission_entities}
\title{Extract Submission-related entities.}
\usage{
extract_submission_entities(
  main,
  questions_df,
  codebook_df,
  base_uri,
  procedure_uri,
  dataset_uri
)
}
\arguments{
\item{main}{The main survey dataframe.}

\item{questions_df}{The questions dataframe.}

\item{codebook_df}{The codebook dataframe.}

\item{base_uri}{The base URI string.}

\item{procedure_uri}{The full IRI of the sur:SurveyProcedure used.}

\item{dataset_uri}{The full IRI of the sur:SurveyDataSet associated with the observations.}
}
\value{
A list of R lists, containing all generated sur:SurveyCompletionTask
and sur:CompletedQuestion entities for all submissions.
}
\description{
This function processes each row of the main dataframe to create
sur:SurveyCompletionTask and associated sur:CompletedQuestion entities.
}
