% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyLevel.R
\name{extract_survey_level_entities}
\alias{extract_survey_level_entities}
\title{Extract top-level entities.}
\usage{
extract_survey_level_entities(main, questions, base_uri, study_id, dataset_id)
}
\arguments{
\item{main}{The main survey data frame.}

\item{questions}{The survey questions data frame.}

\item{base_uri}{The base URI as a string.}

\item{study_id}{A unique identifier string for the overall study.}

\item{dataset_id}{A unique identifier string for the data set.}
}
\value{
A list containing three elements: 'survey_ro', 'procedure', 'dataset'.
}
\description{
This function creates the sur:Survey (Research Object), sur:SurveyProcedure,
and sur:SurveyDataSet entities for a given survey.
}
