% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_html.R
\name{generate_html}
\alias{generate_html}
\title{generate_html}
\usage{
generate_html(entity)
}
\arguments{
\item{entity}{A list representing an entity, typically containing metadata such as \code{rdfs:label}, \code{dc:title}, or \code{sur:hasTitle}.}
}
\value{
A string containing the HTML representation of the entity.
}
\description{
Function to generate an HTML representation of an entity.
}
