% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subdir_from_id.R
\name{get_subdir_from_id}
\alias{get_subdir_from_id}
\title{get_subdir_from_id}
\usage{
get_subdir_from_id(id, idx = 5)
}
\arguments{
\item{id}{a string representing an ID, typically a path or URL}

\item{idx}{an integer representing the index of the subdirectory to extract. default is 5.}
}
\value{
a string representing the subdirectory at the specified index, or "misc" if the index is out of bounds.
}
\description{
get_subdir_from_id
}
\examples{
get_subdir_from_id("a/b/c/d/e/f/g")
}
