% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag-facet.R
\name{get_tags}
\alias{get_tags}
\title{Get labels asigned to each panel of a tagged plot}
\usage{
get_tags(plot = ggplot2::last_plot())
}
\arguments{
\item{plot}{a plot object}
}
\description{
Get labels asigned to each panel of a tagged plot
}
