% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{import_csv_files}
\alias{import_csv_files}
\title{Import a list of CSV files into R.}
\usage{
import_csv_files(path)
}
\arguments{
\item{path}{The absolute path to the .csv files to import.}
}
\description{
Generates a file list of all files of type .csv within a specified path.
These files are then imported using \code{read.csv()}.
}
