% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{load_forms}
\alias{load_forms}
\title{Load a forms tibble from (.Rda) file}
\usage{
load_forms(path)
}
\arguments{
\item{path}{Path to the .Rda forms file to read}
}
\description{
Load a forms tibble from (.Rda) file
}
\examples{
\dontrun{
forms <- load_forms("forms.Rda")
}

}
