% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateHandler.R
\name{load_tag_state}
\alias{load_tag_state}
\title{Load tagging state from disk}
\usage{
load_tag_state(path)
}
\arguments{
\item{path}{Path from where the tag_state should be read}
}
\description{
Load tagging state from disk
}
\examples{
\dontrun{
state <- load_tag_state("question_tagger_state.rds")
}

}
