% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSurveys.R
\name{loadSurveys}
\alias{loadSurveys}
\title{loadSurveys}
\usage{
loadSurveys(path)
}
\arguments{
\item{path}{Character string specifying the path to the survey directories.}
}
\value{
List of survey data frames, where each element corresponds to a survey directory.
}
\description{
loadSurveys
}
\examples{
loadSurveys("path/to/surveys")
}
