% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_tags.R
\name{locate_tags}
\alias{locate_tags}
\title{Locate Part of Speech Tags}
\usage{
locate_tags(x, tags, negate = FALSE, regex = FALSE, logical = TRUE, ...)
}
\arguments{
\item{x}{A \code{tag_pos} object or a named list of vectors.}

\item{tags}{The tags (or regex of \code{regex = TRUE} to select.}

\item{negate}{logical.  If \code{TRUE} and \code{regex = FALSE} the tags not
matching \code{tags} are returned.}

\item{regex}{logical.  If \code{TRUE} then \code{tag} will be grepped for as
a regex.}

\item{logical}{logical.  If \code{TRUE} the out put while be logical \code{TRUE}/\code{FALSE}.
If \code{FALSE} output will be integer element indices.}

\item{\ldots}{Other arguments passed to \code{\link[base]{grepl}}.}
}
\value{
Returns a list of locations (integer or logical) of part of speech
tags as vectors.
}
\description{
Locate part of speech tags within a \code{tag_pos} object.
}
\examples{
data(presidential_debates_2012_pos)
library(dplyr)

presidential_debates_2012_pos \%>\%
    locate_tags(c("NN", "NNP", "NNPS", "NNS"))

presidential_debates_2012_pos \%>\%
    locate_tags(c("NN", "NNP", "NNPS", "NNS"), logical=FALSE)

presidential_debates_2012_pos \%>\%
    locate_tags("NN", regex=TRUE)

presidential_debates_2012_pos \%>\%
    locate_tags("NN(\\\\b|[^S])", regex=TRUE)

presidential_debates_2012_pos \%>\%
    locate_tags(c("NN", "NNP", "NNPS", "NNS"), negate=TRUE)
}
