% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateHandler.R
\name{new_tag_state}
\alias{new_tag_state}
\title{A list to keep track of tagging progress.
The tag_state is saved as progress is made so the pipeline can be resumed in
case it is paused or interrupted.}
\usage{
new_tag_state(questions)
}
\description{
A list to keep track of tagging progress.
The tag_state is saved as progress is made so the pipeline can be resumed in
case it is paused or interrupted.
}
