% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_survey_name.R
\name{normalize_survey_name}
\alias{normalize_survey_name}
\title{normalize_survey_name}
\usage{
normalize_survey_name(name)
}
\arguments{
\item{name}{A character string representing the survey name to be normalized.}
}
\value{
A character string with the normalized survey name in PascalCase.
}
\description{
This function normalizes a survey name by removing leading "X" before a number,
then replacing underscores and punctuation with spaces, and finally
capitalizing each word to create a PascalCase format.
}
\examples{
normalize_survey_name("X1st_survey_data")

}
