% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{ollama_config}
\alias{ollama_config}
\title{Internal config getter}
\usage{
ollama_config(
  base_url = Sys.getenv("OLLAMA_BASE_URL", DEFAULT_OLLAMA_BASE),
  embed_model = Sys.getenv("EMBED_MODEL", DEFAULT_EMBED_MODEL),
  tagger_model = Sys.getenv("TAGGER_MODEL", DEFAULT_TAGGER_MODEL)
)
}
\value{
A list of Ollama configuration parameters.
}
\description{
Invoke this function to return the list of configuration parameters to use
in subsequent embedding or tagging logic.
}
\examples{
ollama_config()
ollama_config(base_url = "http://localhost:11434")

}
