% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penn_tags.R
\name{penn_tags}
\alias{penn_tags}
\title{Penn Treebank Parts of Tags}
\usage{
penn_tags(print = TRUE)
}
\arguments{
\item{print}{logical.  If \code{TRUE} a left justified pretty print version of the
data.frame is printed.}
}
\value{
Invisibly returns a data frame of tags and meaning.
}
\description{
A list of Penn Treebank parts of tags and their meaning.
}
\examples{
penn_tags()
}

