% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_tags.R
\name{plot.count_tags}
\alias{plot.count_tags}
\title{Plots a count_tags object}
\usage{
\method{plot}{count_tags}(x, labels = FALSE, low = "white", high = "red",
  grid = NA, label.color = "grey70", label.size = 3,
  weight = "proportion", ...)
}
\arguments{
\item{x}{The count_tags object.}

\item{labels}{logical.  If \code{TRUE} the cell count values will be included
on the heatmap.}

\item{low}{The color to be used for lower values.}

\item{high}{The color to be used for higher values.}

\item{grid}{The color of the grid (Use \code{NA} to remove the grid).}

\item{label.color}{The color to make labels if \code{labels = TRUE}.}

\item{label.size}{The size to make labels if \code{labels = TRUE}.}

\item{weight}{The weight to apply to the cell values for gradient fill.
Currently the following are available:
\code{"proportion"}, \code{"percent"}.  See \code{\link[termco]{weight}} for
additional information.}

\item{\ldots}{ignored}
}
\description{
Plots a count_tags object.
}

