% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_pos.R
\name{plot.tag_pos}
\alias{plot.tag_pos}
\title{Plots a tag_pos Object}
\usage{
\method{plot}{tag_pos}(x, item.name = "POS Tag", ...)
}
\arguments{
\item{x}{The tag_pos object}

\item{item.name}{The name of the variable that contains the groups (different
element in the vector/list).}

\item{\ldots}{Other arguments passed to \code{\link[termco]{plot_counts}}.}
}
\description{
Plots a tag_pos object.
}
