% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.R
\name{prepare_questions}
\alias{prepare_questions}
\title{Prepare survey questions for tagging}
\usage{
prepare_questions(
  flattened_forms,
  id_col = NULL,
  text_col = "caption",
  n = NULL
)
}
\arguments{
\item{flattened_forms}{description}

\item{id_col}{description}

\item{n}{}
}
\description{
This function prepares questions for tagging by extracting distinct question
captions from the given flattened form tibble.  An optional (integer) slice
value (n) may be specified to select only n questions for tagging (used for demo
purposes to speed up the tagging pipeline).
}
