% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagger-package.R
\docType{data}
\name{presidential_debates_2012_pos}
\alias{presidential_debates_2012_pos}
\title{Parts of Speech for 2012 U.S. Presidential Debates Dialogue}
\format{A list with 2912 elements}
\usage{
data(presidential_debates_2012_pos)
}
\description{
A dataset containing a list of named vectors of words.  The word names are the
part of speech tags.
}
\keyword{datasets}
