% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tuple.R
\name{print.as_tuple}
\alias{print.as_tuple}
\title{Prints a as_tuple Object}
\usage{
\method{print}{as_tuple}(x, truncate = 1500, ...)
}
\arguments{
\item{x}{The as_tuple object.}

\item{truncate}{The number of characters to truncate the output.  Use \code{Inf}
to see all output.}

\item{\ldots}{Other arguments passed to \code{\link[base]{cat}}.  The \code{file}
argument is of particular use as the \code{as_tuple} object is a list of
vectors not as it prints.  Using \code{file} allows the user to export as it prints.}
}
\description{
Prints a as_tuple object.
}
