% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/count_pos.R
\name{print.count_pos}
\alias{print.count_pos}
\title{Prints a count_pos Object}
\usage{
\method{print}{count_pos}(x, digits = 1, weight = "percent",
  zero.replace = "0", pretty = getOption("tagger_pretty"), ...)
}
\arguments{
\item{x}{The count_pos object.}

\item{digits}{The number of digits displayed.}

\item{weight}{The weight type.  Currently the following are available:
\code{"proportion"}, \code{"percent"}.  See \code{\link[termco]{weight}} for
additional information.}

\item{zero.replace}{The value to replace zero count elements with; defaults
to \code{"0"}.}

\item{pretty}{logical.  If \code{TRUE} the counts print in a pretty fashion,
combining count and weighted information into a single display.
\code{pretty} printing can be permanantly removed with
\code{\link[termco]{as_count}}.}

\item{\ldots}{ignored}
}
\description{
Prints a count_pos object.
}

