% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tag_pos.R
\name{print.tag_pos}
\alias{print.tag_pos}
\title{Prints a tag_pos Object}
\usage{
\method{print}{tag_pos}(x, n = 10, width = 0.7 * getOption("width"), ...)
}
\arguments{
\item{x}{The tag_pos object.}

\item{n}{The number of rows to print (.5 * n goes to head and .5 * n goes to tail).}

\item{width}{The width of the tag rows to print.}

\item{\ldots}{ignored}
}
\description{
Prints a tag_pos object.
}

