% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionLevel.R
\name{question_classifier}
\alias{question_classifier}
\title{Classify Questions with Tags}
\usage{
question_classifier(question_text, section_name = NULL)
}
\arguments{
\item{question_text}{The text of the question to classify.}

\item{section_name}{Optional; the name of the section the question belongs to.}
}
\description{
Classify Questions with Tags
}
