% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subdir_from_id.R
\name{sanitize_id}
\alias{sanitize_id}
\title{sanitize_id}
\usage{
sanitize_id(id = NULL)
}
\arguments{
\item{id}{Character string representing an ID, typically a path or URL}
}
\value{
a sanitized string containing only the last segment of the ID.
}
\description{
This function sanitizes an ID by removing any leading path components, leaving only the last segment.
}
\examples{
sanitize_id("a/b/c/d/e/f/g")

}
