% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagging.R
\name{sanitize_label}
\alias{sanitize_label}
\title{Sanitize a raw label into a single word}
\usage{
sanitize_label(raw)
}
\arguments{
\item{raw}{Raw label from the model.}
}
\value{
Sanitized label.
}
\description{
Sanitize a raw label into a single word
}
\keyword{internal}
