% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateHandler.R
\name{save_tag_state}
\alias{save_tag_state}
\title{Save tagging state to disk}
\usage{
save_tag_state(state, path)
}
\arguments{
\item{state}{tag_state to write to disk}

\item{path}{Path where the tag_state should be written}
}
\description{
Save tagging state to disk
}
\examples{
state <- list(questions = tibble::tibble(id = "q_001", caption = "Age?"))
class(state) <- "tag_state"
tmp <- tempfile(fileext = ".rds")
save_tag_state(state, tmp)

}
