% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/select_tags.R
\name{select_tags}
\alias{select_tags}
\title{Select Tags + Words From tag_pos Object}
\usage{
select_tags(x, tags, negate = FALSE, regex = FALSE, ...)
}
\arguments{
\item{x}{A \code{tag_pos} object or a named list of vectors.}

\item{tags}{The tags (or regex of \code{regex = TRUE} to select.}

\item{negate}{logical.  If \code{TRUE} and \code{regex = FALSE} the tags not
matching \code{tags} are returned.}

\item{regex}{logical.  If \code{TRUE} then \code{tag} will be grepped for as
a regex.}

\item{\ldots}{Other arguments passed to \code{\link[base]{grepl}}.}
}
\value{
Returns a list of selected part of speech tagged vectors.
}
\description{
Select part of speech tags and accompanying words.
}
\examples{
data(presidential_debates_2012_pos)
library(dplyr)

presidential_debates_2012_pos \%>\%
    select_tags(c("NN", "NNP", "NNPS", "NNS"))

presidential_debates_2012_pos \%>\%
    select_tags(c("NN", "NNP", "NNPS", "NNS")) \%>\%
    count_tags(presidential_debates_2012[, c("person", "time")])

presidential_debates_2012_pos \%>\%
    select_tags("NN", regex=TRUE) \%>\%
    count_tags(presidential_debates_2012[, c("person", "time")])

presidential_debates_2012_pos \%>\%
    select_tags("NN(\\\\b|[^S])", regex=TRUE) \%>\%
    count_tags(presidential_debates_2012[, c("person", "time")])

presidential_debates_2012_pos \%>\%
    select_tags(c("NN", "NNP", "NNPS", "NNS"), TRUE) \%>\%
    count_tags(presidential_debates_2012[, c("person", "time")])
}

