% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSurveyContext.R
\name{set_survey_context}
\alias{set_survey_context}
\title{Set the survey context.}
\usage{
set_survey_context(additional_context = NULL)
}
\arguments{
\item{additional_context}{Any additional context fields to add to the base
context; given as a \code{list}.}
}
\value{
A list representing the context entity.
}
\description{
Set the context to use when capturing the data for a given survey.
This context set can be updated accordingly, but as a start the primary
ontologies referenced are given.
}
