% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama.R
\name{test_connection}
\alias{test_connection}
\title{Test Ollama connection}
\usage{
test_connection(base_url, model_name)
}
\arguments{
\item{base_url}{Base url where the Ollama instance is running.}

\item{model_name}{Desired model to use for embedding/tagging.}
}
\value{
TRUE on connection success.
}
\description{
Before running the tagging pipeline configure and test the Ollama connection
to ensure the necessary models are running and reachable.
}
\examples{
\dontrun{
test_connection("http://localhost:11434", "llama3.1:8b")
}

}
