% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_pos.R
\name{tidy_pos}
\alias{tidy_pos}
\title{Tidy Parts of Speech Object}
\usage{
tidy_pos(x, id.name = "id", token.name = "token", pos.name = "pos", ...)
}
\arguments{
\item{x}{A \code{tag_pos} object.}

\item{id.name}{The name for the column assigned to each element in the text vector.}

\item{token.name}{The name for the column of tokens.}

\item{pos.name}{The name for the column of parts of speech.}

\item{\ldots}{ignored.}
}
\value{
Returns a \code{\link[data.table]{data.table}} with element, token,
and pos columns.
}
\description{
Pull out text element id, tokens, and parts of speech as separate columns in a
tidy data.frame.
}
\examples{
(out1 <- tag_pos(sam_i_am))
tidy_pos(out1)
}

