% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagger-package.R
\docType{data}
\name{universal_pos_map}
\alias{universal_pos_map}
\title{Mapping to Universal Tags}
\format{A list with 33 elements}
\usage{
data(universal_pos_map)
}
\description{
A dataset containing a list \code{\link[base]{data.frame}} maps to move from
Penn Treebank tagsets to a set of 12 universal part of speech tags.  This
mapping comes from Petrov, Das, & McDonald (2011).
}
\details{
Each \code{\link[base]{data.frame}} contains a tag column,
corresponding to the Penn Treebank tagset, and a universal column,
corresponding to the 12 universal tags.
}
\references{
\url{https://github.com/slavpetrov/universal-pos-tags}

Slav Petrov, Dipanjan Das and Ryan McDonald. (2011). A Universal Part-of-Speech Tagset. http://arxiv.org/abs/1104.2086
}
\keyword{datasets}
