% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_entity_files.R
\name{write_entity_files}
\alias{write_entity_files}
\title{write_entity_files}
\usage{
write_entity_files(entity = NULL, gh_pages_path = NULL, subdir = NULL)
}
\arguments{
\item{entity}{A list representing the entity to write, typically in JSON-LD format.}
}
\value{
Writes JSON-LD and HTML files for the entity to the specified GitHub Pages path.
}
\description{
Function to write jsonld and html files for an entity to the appropriate subdirectory
in the GitHub Pages repository idenfied by gh_pages_path and subdir.
}
\examples{
walk(all_graph_entities, write_entity_files)
}
