% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as_tuple.R
\name{as_tuple}
\alias{as_tuple}
\title{Combine Words + Tags Tuple}
\usage{
as_tuple(x, word.first = TRUE, ...)
}
\arguments{
\item{x}{A \code{tag_pos} object or a named list of vectors.}

\item{word.first}{logical.  If \code{TRUE} the word comes before the tag.}

\item{\ldots}{ignored.}
}
\value{
Returns a lst of lists of tuples vectors..
}
\description{
A traditional tuple style of displaying words and tags in the same vector.
}
\examples{
(x <- tag_pos("I need $54 to go to the movies."))
c(x) ## The true structure of a `tag_pos` object
as_tuple(c(x))

## Don't truncate the printing
print(as_tuple(c(x)), truncate=FALSE)

## Export tuple printing to file
\dontrun{
print(as_tuple(c(x)), truncate=Inf, file="out.txt")
}
as_tuple(c(x), word.first=FALSE)

(out1 <- tag_pos(sam_i_am))
as_tuple(c(out1))
}

